Exponential backoff in SW
Aaaaand here I ran out of time. I was trying to implement an exponential backoff for the api, so if it does get rate limited it would try and store the requests for retrying. The idea was that if the user stays on the page, the retry would fire and then postMessage back to the window with updated date. Changing page would clear out the queue (so we don't make unnecessary requests).